requirements.txt
Django==3.1.3
psycopg2-binary

django-cors-headers==3.5.0
djangorestframework==3.12.2
gunicorn==20.0.4
whitenoise==5.2.0
Pillow
PyMySQL